+++ src/EditorFeatures/Core/Tagging/AsynchronousViewportTaggerProvider.cs
@@ -2,24 +2,18 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-using System.Collections.Generic;
 using System.Collections.Immutable;
-using System.Threading.Tasks;
 using System.Threading;
-using EnvDTE80;
-using Microsoft.CodeAnalysis.Editor.Shared.Extensions;
+using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Editor.Shared.Utilities;
 using Microsoft.CodeAnalysis.Options;
 using Microsoft.CodeAnalysis.Shared.Collections;
 using Microsoft.CodeAnalysis.Shared.TestHooks;
+using Microsoft.CodeAnalysis.Tagging;
 using Microsoft.CodeAnalysis.Workspaces;
-using Microsoft.VisualStudio.Shell.Interop;
 using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Text.Editor;
 using Microsoft.VisualStudio.Text.Tagging;
-using Roslyn.Utilities;
-using Microsoft.CodeAnalysis.Tagging;
 
 namespace Microsoft.CodeAnalysis.Editor.Tagging;
 
@@ -29,101 +23,13 @@ namespace Microsoft.CodeAnalysis.Editor.Tagging;
 /// cref="AbstractAsynchronousTaggerProvider{TTag}.EventChangeDelay"/>, but will also compute tags for the regions
 /// above and below that visible section at a delay of <see cref="DelayTimeSpan.NonFocus"/>.
 /// </summary>
-internal abstract class AsynchronousViewportTaggerProvider<TTag> : IViewTaggerProvider
+internal abstract partial class AsynchronousViewportTaggerProvider<TTag> : IViewTaggerProvider
     where TTag : ITag
 {
     private enum ViewPortToTag
     {
-        AboveAndBelow,
         InView,
-    }
-
-    private sealed class SingleViewportTaggerProvider : AsynchronousViewTaggerProvider<TTag>
-    {
-        private readonly AsynchronousViewportTaggerProvider<TTag> _callback;
-
-        private readonly ViewPortToTag _viewPortToTag;
-
-        public SingleViewportTaggerProvider(
-            AsynchronousViewportTaggerProvider<TTag> callback,
-            ViewPortToTag viewPortToTag,
-            IThreadingContext threadingContext,
-            IGlobalOptionService globalOptions,
-            ITextBufferVisibilityTracker? visibilityTracker,
-            IAsynchronousOperationListener asyncListener)
-            : base(threadingContext, globalOptions, visibilityTracker, asyncListener)
-        {
-            _callback = callback;
-            _viewPortToTag = viewPortToTag;
-        }
-
-        protected override ImmutableArray<IOption2> Options
-            => _callback.Options;
-
-        protected override TaggerTextChangeBehavior TextChangeBehavior
-            => _callback.TextChangeBehavior;
-
-        protected override ITaggerEventSource CreateEventSource(ITextView textView, ITextBuffer subjectBuffer)
-            => _callback.CreateEventSource(textView, subjectBuffer);
-
-        protected override TaggerDelay EventChangeDelay
-            => _viewPortToTag == ViewPortToTag.InView ? _callback.EventChangeDelay : TaggerDelay.NonFocus;
-
-        protected override IEnumerable<SnapshotSpan> GetSpansToTag(ITextView? textView, ITextBuffer subjectBuffer)
-        {
-            this.ThreadingContext.ThrowIfNotOnUIThread();
-            Contract.ThrowIfNull(textView);
-
-            // if we're the current view, attempt to just get what's visible, plus 10 lines above and below.  This will
-            // ensure that moving up/down a few lines tends to have immediate accurate results.
-            var visibleSpanOpt = textView.GetVisibleLinesSpan(subjectBuffer, extraLines: 10);
-            if (visibleSpanOpt is null)
-            {
-                // couldn't figure out the visible span.  So the InView tagger will need to tag everything, and the
-                // above/below tagger should tag nothing.
-                return _viewPortToTag == ViewPortToTag.InView
-                    ? base.GetSpansToTag(textView, subjectBuffer)
-                    : SpecializedCollections.EmptyEnumerable<SnapshotSpan>();
-            }
-
-            var visibleSpan = visibleSpanOpt.Value;
-
-            // If we're the 'InView' tagger, tag what was visible. 
-            if (_viewPortToTag is ViewPortToTag.InView)
-                return SpecializedCollections.SingletonEnumerable(visibleSpan);
-
-            // For the above/below tagger, broaden the span to to the requested portion above/below what's visible, then
-            // subtract out the visible range.
-            var widenedSpanOpt = textView.GetVisibleLinesSpan(subjectBuffer, extraLines: _callback._extraLinesAroundViewportToTag);
-            Contract.ThrowIfNull(widenedSpanOpt, "Should not ever fail getting the widened span as we were able to get the normal span");
-
-            var widenedSpan = widenedSpanOpt.Value;
-            Contract.ThrowIfFalse(widenedSpan.Span.Contains(visibleSpan.Span), "The widened span must be at least as large as the visible one.");
-
-            using var result = TemporaryArray<SnapshotSpan>.Empty;
-
-            var aboveSpan = new SnapshotSpan(visibleSpan.Snapshot, Span.FromBounds(widenedSpan.Span.Start, visibleSpan.Span.Start));
-            var belowSpan = new SnapshotSpan(visibleSpan.Snapshot, Span.FromBounds(visibleSpan.Span.End, widenedSpan.Span.End));
-
-            if (!aboveSpan.IsEmpty)
-                result.Add(aboveSpan);
-
-            if (!belowSpan.IsEmpty)
-                result.Add(belowSpan);
-
-            return result.ToImmutableAndClear();
-        }
-
-        protected override async Task ProduceTagsAsync(
-            TaggerContext<TTag> context, CancellationToken cancellationToken)
-        {
-            foreach (var spanToTag in context.SpansToTag)
-            {
-                cancellationToken.ThrowIfCancellationRequested();
-                await _callback.ProduceTagsAsync(
-                    context, spanToTag, cancellationToken).ConfigureAwait(false);
-            }
-        }
+        AboveAndBelow,
     }
 
     /// <summary>
@@ -169,7 +75,7 @@ internal abstract class AsynchronousViewportTaggerProvider<TTag> : IViewTaggerPr
     }
 
     // Functionality for subclasses to control how this diagnostic tagging operates.  All the individual
-    // SingleDiagnosticKindTaggerProvider will defer to these to do the work so that they otherwise operate
+    // SingleViewportTaggerProvider will defer to these to do the work so that they otherwise operate
     // identically.
 
     /// <inheritdoc cref="AbstractAsynchronousTaggerProvider{TTag}.Options"/>
@@ -177,11 +83,6 @@ internal abstract class AsynchronousViewportTaggerProvider<TTag> : IViewTaggerPr
 
     /// <inheritdoc cref="AbstractAsynchronousTaggerProvider{TTag}.TextChangeBehavior"/>
     protected virtual TaggerTextChangeBehavior TextChangeBehavior => TaggerTextChangeBehavior.None;
-
-    //protected virtual ImmutableArray<IOption2> FeatureOptions { get; } = ImmutableArray<IOption2>.Empty;
-
-    //protected abstract bool TagEquals(TTag tag1, TTag tag2);
-    // protected abstract ITagSpan<TTag>? CreateTagSpan(Workspace workspace, bool isLiveUpdate, SnapshotSpan span, DiagnosticData data);
 
     /// <inheritdoc cref="AbstractAsynchronousTaggerProvider{TTag}.CreateEventSource(ITextView?, ITextBuffer)"/>
     protected abstract ITaggerEventSource CreateEventSource(ITextView textView, ITextBuffer subjectBuffer);
